requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0,<2.0.0
pandas>=1.5.0
pillow>=9.5.0
scikit-learn>=1.3.0
plotly>=5.15.0 